+++ common/src/main/java/net/mehvahdjukaar/vista/VistaMod.java
@@ -177,9 +177,12 @@ public class VistaMod {
         }
 
 
+        //TODO: improve watermedia
+        //TODO: fix channel 2 being blue
+        //TODO: fix pause not working
         //TODO: alex mobs shaders compat
-        //TODO:
-        //tinted panes
+        //TODO: check clock overlay
+        //TODO: check panes not working for some
         //night vision and ctr effects for view finders
         //better way to get a clock overlay on tvs
         //check update rate when multiple view finders are queried