+++ common/src/main/java/net/mehvahdjukaar/vista/client/renderer/TvBlockEntityRenderer.java
@@ -106,7 +106,7 @@ public class TvBlockEntityRenderer implements BlockEntityRenderer<TVBlockEntity>
 
         IVideoSource videoSource = blockEntity.getVideoSource();
         VertexConsumer vc = videoSource
-                .getVideoFrameBuilder(partialTick, buffer,
+                .getVideoFrameBuilder(paused ? 1 : partialTick, buffer,
                         shouldUpdate, screenSize, pixelEffectRes,
                         videoTicks, paused, switchAnim, staticAnim);