+++ common/src/main/java/net/mehvahdjukaar/vista/integration/watermedia/WatermediaSession.java
@@ -2,6 +2,7 @@ package net.mehvahdjukaar.vista.integration.watermedia;
 
 import net.mehvahdjukaar.vista.client.textures.IWebTexture;
 import net.mehvahdjukaar.vista.client.web.IMediaSession;
+import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
 import org.watermedia.api.image.ImageAPI;
 import org.watermedia.api.image.ImageCache;
@@ -19,7 +20,8 @@ public class WatermediaSession implements IMediaSession {
 
     public WatermediaSession(String url, Executor executor,
                              int targetWidth, int targetHeight) {
-        this.imageCache = ImageAPI.getCache(URI.create(url.trim()), executor);
+        //TODO: check executor
+        this.imageCache = ImageAPI.getCache(URI.create(url.trim()), Minecraft.getInstance());
         this.executor = executor;
         this.targetWidth = targetWidth;
         this.targetHeight = targetHeight;
@@ -28,13 +30,13 @@ public class WatermediaSession implements IMediaSession {
 
     @Override
     public IWebTexture createTextureView(ResourceLocation resourceLocation) {
-        return true ?
+        return (true || imageCache.getException() != null || imageCache.isVideo() || imageCache.getRenderer() == null) ?
                 new WatermediaVideoTexture(resourceLocation, imageCache, targetWidth, targetHeight, executor) :
                 new WatermediaImageTexture(resourceLocation, imageCache, targetWidth, targetHeight, executor);
     }
 
     @Override
     public void close() {
-        imageCache.deuse();
+        //imageCache.deuse();
     }
 }